requirements.txt
streamlit 
langchain_community
langchain_openai
langchain_core
python-dotenv
langserve
fastapi 
uvicorn
sse_starlette